feat: add MCP server, parameters system, forge orchestration, spatial-llm, and infrastructure updates
- Add MCP crate with EEP v1.0 protocol, HTTP endpoints, and export targets
- Add parameters module with GlobalParameters, DomainRegistry, InstanceParameters
- Add forge jobs, routing, and updated config for forge-orchestration 0.4.x
- Add spatial-llm client, context, generation, indexing, prompt, query modules
- Add embedvec crate for vector embeddings
- Add orbital navigation and USD native format architecture docs
- Add EEP specification documentation
- Add infrastructure: Nomad jobs, Terraform configs, forge scripts
- Update common crate: attributes, generation, orbital, scene_ops, services
- Update web app pages and styling
- Add WATER directory

+++ eustress/crates/common/src/classes.rs
@@ -6709,7 +6709,7 @@ impl RegionChunk {
             crate::orbital::Region::from_geodetic(
                 self.center_geodetic[0],
                 self.center_geodetic[1],
-                self.bounds_extents.x * 2.0,
+                (self.bounds_extents.x * 2.0) as f64,
             )
         }
     }
